+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Change GTM push of unset values from 'null' to 'undefined' ([PR #2971](https://github.com/alphagov/govuk_publishing_components/pull/2971))
+* Add click tracking to Yes/No buttons on feedback component ([PR #2964](https://github.com/alphagov/govuk_publishing_components/pull/2964))
 
 ## 30.6.1